+++ R/Get_Ibov_Stocks.R
@@ -2,6 +2,7 @@
 #'
 #' This function scrapes the stocks that constitute tSP500 index from the wikipedia page at http://bvmf.bmfbovespa.com.br/indices/ResumoCarteiraTeorica.aspx?Indice=IBOV&idioma=pt-br.
 #'
+#' @param max.tries Maximum number of attempts to download the data
 #' @return A dataframe that includes a column with the list of tickers of companies that belong to the Ibovespa index
 #' @export
 #' @examples
@@ -9,12 +10,20 @@
 #' df.ibov <- GetIbovStocks()
 #' print(df.ibov$tickers)
 #' }
-GetIbovStocks <- function(){
+GetIbovStocks <- function(max.tries  = 10){
 
   # get list of ibovespa's tickers from wbsite
 
-  myUrl <- 'http://bvmf.bmfbovespa.com.br/indices/ResumoCarteiraTeorica.aspx?Indice=IBOV&idioma=pt-br'
-  df.ibov.comp <- XML::readHTMLTable(myUrl)[[1]]
+  for (i.try in seq(max.tries)) {
+    myUrl <- 'http://bvmf.bmfbovespa.com.br/indices/ResumoCarteiraTeorica.aspx?Indice=IBOV&idioma=pt-br'
+    #df.ibov.comp <- XML::readHTMLTable(myUrl)[[1]]
+    df.ibov.comp <- as.data.frame(XML::readHTMLTable(myUrl))
+
+    Sys.sleep(0.5)
+
+    if (nrow(df.ibov.comp) > 0) break()
+
+  }
 
   names(df.ibov.comp) <- c('tickers', 'ticker.desc', 'type.stock', 'quantity', 'percentage.participation')